gc(wasm): prune WASM_MEMORY_BINDINGS entries whose buffer died (#8174 custody), register it in DEAD_KEY_PRUNES + the rekey manifest; split wasm-host tables for the file cap
