+++ scripts/gc_rekeyed_key_tables.json
@@ -56,6 +56,12 @@
       "why": "Registered prune retains on !is_dead_owner over the iterator array address (map.rs:54)."
     },
     {
+      "site": "crates/perry-runtime/src/node_api_host/metadata.rs::scan_object_metadata_roots_mut",
+      "table": "NODE_API_OBJECT_METADATA",
+      "death": "dead_owner:prune_dead_object_meta_owners",
+      "why": "Node-API object metadata is re-keyed without marking its owner; the registered prune removes dead keys after both copying and mark/sweep traces and queues their finalizers on the owning agent."
+    },
+    {
       "site": "crates/perry-runtime/src/node_submodules/diagnostics.rs::scan_diagnostics_channel_key_roots_mut",
       "table": "DIAG_CHANNEL_BY_KEY",
       "death": "key_is_rooted",
@@ -242,10 +248,10 @@
       "why": "Called directly from prune_dead_owner_side_tables_post_trace (gc/dead_owner.rs), not from fan_out; the copying path prunes inline via HolderDisposition::Drop in process_weak_targets_from_registry (weakref.rs:960-966)."
     },
     {
-      "site": "crates/perry-runtime/src/node_api_host/metadata.rs::scan_object_metadata_roots_mut",
-      "table": "NODE_API_OBJECT_METADATA",
-      "death": "dead_owner:prune_dead_object_meta_owners",
-      "why": "Node-API object metadata is re-keyed without marking its owner; the registered prune removes dead keys after both copying and mark/sweep traces and queues their finalizers on the owning agent."
+      "site": "crates/perry-runtime/src/webassembly.rs::scan_wasm_memory_binding_roots_mut",
+      "table": "WASM_MEMORY_BINDINGS",
+      "death": "dead_owner:prune_dead_wasm_memory_bindings",
+      "why": "The key is a native wasmi instance pointer, so the GC address at risk is the VALUE (MemoryBinding.buffer). Registered prune retains on !is_dead_owner over that buffer address (webassembly.rs), so a Memory swept after its JS owner became unreachable takes its entry with it instead of leaving an address for visit_metadata_usize_slot to re-point once it is recycled and forwarded."
     }
   ]
 }
